# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Authn Platform API
  description: Enterprise Identity Engine REST API specification for authentication, session management, organization multi-tenancy,
    RBAC, SAML, OAuth2/OIDC, Webhooks, API key administration, and account recovery.
  version: 2.0.0
  contact:
    name: Authn Platform Engineering
    url: https://github.com/hanan-bhatti/authn-2.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
- url: http://localhost:8080
  description: Local Development Engine
- url: https://api.staging.authn.dev
  description: Staging Environment
- url: https://api.authn.dev
  description: Production Environment
paths:
...
  /v1/admin/jwks/rotate:
    post:
      summary: Rotate JWKS Signing Key
      description: Promotes a freshly generated RSA key to active and moves the outgoing key into the rotation grace period.
        Returns the updated key set.
      operationId: rotateJWKS
      tags:
      - Key Management
      security:
      - SecretKey: []
      - BearerAuth: []
      responses:
        '200':
          description: JWKS key rotated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RotateJWKSResponse'
        '401':
          description: Unauthorized - Secret key or console JWT missing/invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden - Tenant admin role or 2FA enrollment required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/tenant/applications:
    post:
      summary: Register Client Application
      description: Registers an OAuth client application under the authenticated tenant with exact redirect URIs allowlist.
      operationId: createApplication
      tags:
      - Application Management
      security:
      - SecretKey: []
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '201':
          description: Application registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApplicationResponse'
        '400':
          description: Bad Request - Missing id or name, or invalid body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error: id and name are required
                code: missing_parameter
        '401':
          description: Unauthorized - Tenant resolution failed or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden - Tenant admin role or 2FA enrollment required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    PublishableKey:
      type: apiKey
      name: X-Authn-Publishable-Key
      in: header
      description: Client publishable key identifying application context
    SecretKey:
      type: apiKey
      name: X-Authn-Secret-Key
      in: header
      description: Backend secret key for tenant admin API calls
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Client user or console admin session JWT
  schemas:
...
    ErrorEnvelope:
      type: object
      required:
      - error
      - code
      properties:
        error:
          type: string
          description: Human-readable error message (never discloses internal implementation details)
          example: invalid or expired verification token
        code:
          type: string
          description: Stable machine-readable error identifier
          example: invalid_token
...
    JSONWebKey:
      type: object
      description: Single RFC 7517 JSON Web Key entry
      required:
      - kty
      - use
      - alg
      - kid
      properties:
        kty:
          type: string
          example: RSA
        use:
          type: string
          example: sig
        alg:
          type: string
          example: RS256
        kid:
          type: string
          example: key_2026_active
        n:
          type: string
          description: Base64url-encoded RSA public modulus
        e:
          type: string
          description: Base64url-encoded RSA public exponent
    JWKSResponse:
      type: object
      description: Document returned by JWKS endpoint
      required:
      - keys
      properties:
        keys:
          type: array
          items:
            $ref: '#/components/schemas/JSONWebKey'
...
    RotateJWKSResponse:
      type: object
      required:
      - message
      - new_active_key_id
      - jwks
      properties:
        message:
          type: string
          example: JWKS key successfully rotated
        new_active_key_id:
          type: string
          example: key_2026_active_v2
        jwks:
          $ref: '#/components/schemas/JWKSResponse'
    CreateApplicationRequest:
      type: object
      required:
      - id
      - name
      properties:
        id:
          type: string
          example: app_my_client
        name:
          type: string
          example: My OAuth Client Application
        exact_redirect_uris:
          type: array
          items:
            type: string
          example:
          - https://app.example.com/callback
    CreateApplicationResponse:
      type: object
      required:
      - message
      - id
      - name
      - exact_redirect_uris
      properties:
        message:
          type: string
          example: application registered successfully
        id:
          type: string
          example: app_my_client
        name:
          type: string
          example: My OAuth Client Application
        exact_redirect_uris:
          type: array
          items:
            type: string
          example:
          - https://app.example.com/callback
...